49 fix statistics layout

+++ src/app/modules/statistics/page/statistics/statistics.container.ts
@@ -1,9 +1,11 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
+import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
 import { id } from '@swimlane/ngx-charts';
 import { Observable, Subject, Subscription } from 'rxjs';
 import { StockTradeBoardService } from 'src/app/core/services/stock-trade-board.service';
 import { StatisticTypeMarker } from 'src/app/data/enums/statistics-markers.enum';
 import { StockSellModel, TransactionWalletModel } from 'src/app/data/models/statistics-section.model';
+import { GlobalDialogComponent } from 'src/app/shared/components/global-dialog/global-dialog.component';
 
 @Component({
     selector: 'app-statistics',
@@ -18,22 +20,31 @@ export class StatisticsContainerComponent implements OnInit, OnDestroy {
     private loseProfitDataMarker: string = this.chosenTypeOfDataForDisplay.profit
 
     private chartColorPallet = {
+
         domain: [
+
             '#00A8FF'
+
         ]
+
     };
 
     private transactionsSubscription: Subscription;
     private transactionsData: StockSellModel[];
 
     private transactionsDataSubject = new Subject<StockSellModel[]>();
-    private transactionsDataSubject$: Observable<StockSellModel[]> = this.transactionsDataSubject.asObservable();
+    private transactionsDataSubject$: Observable<StockSellModel[]> =
+        this.transactionsDataSubject.asObservable();
+
+    private transactionsDataProfitLoseSubject = new Subject<TransactionWalletModel[]>();
+    private transactionsDataProfitLoseSubject$: Observable<TransactionWalletModel[]>
+        = this.transactionsDataProfitLoseSubject.asObservable();
 
     private profitLossesData: any;
     private linearChartData: any;
     private transactionWallet: TransactionWalletModel[];
 
-    constructor(private stockTradeBoardService: StockTradeBoardService) { }
+    constructor(private stockTradeBoardService: StockTradeBoardService, public matDialog: MatDialog) { }
 
     ngOnInit(): void {
 
@@ -64,7 +75,7 @@ export class StatisticsContainerComponent implements OnInit, OnDestroy {
     //==========================================================================
 
     /**
-     * 
+     *  total lost nad profit data
      */
     get getProfitLossesData(): any {
 
@@ -73,7 +84,16 @@ export class StatisticsContainerComponent implements OnInit, OnDestroy {
     }
 
     /**
-   * 
+     * transaction wallet as an observable
+     */
+    get getProfitLossesData$(): Observable<TransactionWalletModel[]> {
+
+        return this.transactionsDataProfitLoseSubject$;
+
+    }
+
+    /**
+   * linear chart data of total profit/los balance
    */
     get getLinearChartData(): any {
 
@@ -82,20 +102,20 @@ export class StatisticsContainerComponent implements OnInit, OnDestroy {
     }
 
     /**
-    * 
-    */
-    get getTransactionWallet(): TransactionWalletModel[] {
+     * transaction data for table display 
+     */
+    get getTransactionsData(): StockSellModel[] {
 
-        return this.transactionWallet;
+        return this.transactionsData;
 
     }
 
     /**
      * 
      */
-    get getTransactionsData(): StockSellModel[] {
+    get getTransactionWallet(): TransactionWalletModel[] {
 
-        return this.transactionsData;
+        return this.transactionWallet;
 
     }
 
@@ -134,8 +154,8 @@ export class StatisticsContainerComponent implements OnInit, OnDestroy {
 
         this.linearChartData = this.calculateLinearChart(this.transactionsData);
 
-        this.transactionWallet = this.generateTransactionsWallet(this.transactionsData);
-
+        this.generateTransactionsWallet(this.transactionsData);
+     
     }
 
     /**
@@ -174,6 +194,8 @@ export class StatisticsContainerComponent implements OnInit, OnDestroy {
         return dataArray;
 
     }
+
+    //==========================================================================
 
     /**
      * Calculating total profits value, total loses value and total trade balance
@@ -248,7 +270,11 @@ export class StatisticsContainerComponent implements OnInit, OnDestroy {
                 });
             }
 
-        });
+        });    
+        
+        this.transactionWallet = dataArray;
+
+        this.transactionsDataProfitLoseSubject.next(dataArray);           
 
         return dataArray;
 
@@ -318,7 +344,7 @@ export class StatisticsContainerComponent implements OnInit, OnDestroy {
      */
     switchProfitLoseCharts(marker: string): void {
 
-        if( this.loseProfitDataMarker === marker) return;
+        if (this.loseProfitDataMarker === marker) return;
 
         //overWriting the marker
         this.loseProfitDataMarker = marker;
@@ -331,6 +357,7 @@ export class StatisticsContainerComponent implements OnInit, OnDestroy {
                 domain: [
 
                     '#00A8FF'
+
                 ]
 
             };
@@ -342,6 +369,7 @@ export class StatisticsContainerComponent implements OnInit, OnDestroy {
                 domain: [
 
                     '#f30625'
+
                 ]
 
             };
@@ -349,17 +377,42 @@ export class StatisticsContainerComponent implements OnInit, OnDestroy {
         }
 
         // generating new list of trades profit or lose
-        this.transactionWallet = this.generateTransactionsWallet(this.transactionsData);
+        this.generateTransactionsWallet(this.transactionsData);
 
     }
+
+    //==========================================================================
 
     /**
      * 
      * @param id 
      */
     deletePositionFromTable(id: string): void {
+        
+        const dialogConfig = new MatDialogConfig();
 
-        this.stockTradeBoardService.deleteTransaction(id);
+        dialogConfig.disableClose = false;
+        dialogConfig.id = "modal-component";
+    
+        dialogConfig.data = {
+          header: 'home.sideNavDeleteTransactionsButton',
+          description: 'home.sideNavDeleteTransactionsDialogText'
+        }
+    
+        // Initializing dialog
+        const modalDialog = this.matDialog
+          .open(GlobalDialogComponent, dialogConfig);
+    
+        // Receive data from dialog
+        modalDialog.afterClosed().subscribe(result => {
+    
+          if (result) {
+    
+            this.stockTradeBoardService.deleteTransaction(id);
+    
+          }
+    
+        });       
 
     }
